Refactor basis vector search strategies to individual modules
