+++ algorithms/indexing/basis_vector_search/test_combinations.py
@@ -8,13 +8,12 @@ import scitbx.matrix
 from scitbx.math import euler_angles_as_matrix
 from dxtbx.model import Crystal
 
-from dials.algorithms.indexing.basis_vector_search import strategies
-from dials.algorithms.indexing.basis_vector_search import combinations
+from . import combinations, fft1d
 
 
 def test_combinations(setup_rlp):
     max_cell = 1.3 * max(setup_rlp["crystal_symmetry"].unit_cell().parameters()[:3])
-    strategy = strategies.FFT1D(max_cell)
+    strategy = fft1d.FFT1D(max_cell)
     basis_vectors, used = strategy.find_basis_vectors(setup_rlp["rlp"])
 
     for target_symmetry in (
