+++ algorithms/indexing/basis_vector_search/test_optimise.py
@@ -3,14 +3,14 @@ from __future__ import absolute_import, division, print_function
 import pytest
 import random
 
-from dials.algorithms.indexing.basis_vector_search import optimise, strategies
+from . import fft3d, optimise
 
 
 def test_optimise_basis_vectors(setup_rlp):
     random.seed(42)  # guaranteed to be random
     max_cell = 1.3 * max(setup_rlp["crystal_symmetry"].unit_cell().parameters()[:3])
     rlp = setup_rlp["rlp"]
-    strategy = strategies.FFT3D(max_cell, n_points=256)
+    strategy = fft3d.FFT3D(max_cell, n_points=256)
     basis_vectors, used = strategy.find_basis_vectors(rlp)
 
     for v in basis_vectors:
